+++ clients/amqp-connection/src/main/java/org/eclipse/hono/client/amqp/connection/impl/HonoConnectionImpl.java
@@ -37,12 +37,12 @@ import org.eclipse.hono.client.ClientErrorException;
 import org.eclipse.hono.client.ServerErrorException;
 import org.eclipse.hono.client.ServiceInvocationException;
 import org.eclipse.hono.client.amqp.config.ClientConfigProperties;
+import org.eclipse.hono.client.amqp.connection.AmqpConstants;
 import org.eclipse.hono.client.amqp.connection.ConnectionFactory;
 import org.eclipse.hono.client.amqp.connection.DisconnectListener;
 import org.eclipse.hono.client.amqp.connection.ErrorConverter;
 import org.eclipse.hono.client.amqp.connection.HonoConnection;
 import org.eclipse.hono.client.amqp.connection.ReconnectListener;
-import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.HonoProtonHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -575,7 +575,7 @@ public final class HonoConnectionImpl implements HonoConnection {
         return executeOnContext(result -> {
             checkConnected().compose(v -> {
 
-                if (targetAddress == null && !supportsCapability(Constants.CAP_ANONYMOUS_RELAY)) {
+                if (targetAddress == null && !supportsCapability(AmqpConstants.CAP_ANONYMOUS_RELAY)) {
                     // AnonTerm spec requires peer to offer ANONYMOUS-RELAY capability
                     // before a client can use anonymous terminus
                     return Future.failedFuture(new ServerErrorException(