+++ service-base/src/test/java/org/eclipse/hono/service/amqp/AbstractRequestResponseEndpointTest.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -34,6 +34,7 @@ import org.apache.qpid.proton.amqp.transport.DeliveryState;
 import org.apache.qpid.proton.message.Message;
 import org.eclipse.hono.auth.HonoUser;
 import org.eclipse.hono.client.ClientErrorException;
+import org.eclipse.hono.client.amqp.connection.AmqpConstants;
 import org.eclipse.hono.config.ServiceConfigProperties;
 import org.eclipse.hono.service.auth.AuthorizationService;
 import org.eclipse.hono.util.Constants;
@@ -231,7 +232,7 @@ public class AbstractRequestResponseEndpointTest {
         verify(delivery).disposition(deliveryState.capture(), eq(Boolean.TRUE));
         assertThat(deliveryState.getValue()).isInstanceOf(Accepted.class);
         verify(receiver, never()).close();
-        verify(authService).isAuthorized(Constants.PRINCIPAL_ANONYMOUS, resource, "unauthorized");
+        verify(authService).isAuthorized(AmqpConstants.PRINCIPAL_ANONYMOUS, resource, "unauthorized");
         // but not being processed
         verify(requestMessageHandler, never()).apply(any(Message.class), any(ResourceIdentifier.class));
         // and a response is sent to the client with status 403
@@ -312,7 +313,7 @@ public class AbstractRequestResponseEndpointTest {
         verify(delivery).disposition(deliveryState.capture(), eq(Boolean.TRUE));
         assertThat(deliveryState.getValue()).isInstanceOf(Accepted.class);
         verify(receiver, never()).close();
-        verify(authService).isAuthorized(Constants.PRINCIPAL_ANONYMOUS, resource, "get");
+        verify(authService).isAuthorized(AmqpConstants.PRINCIPAL_ANONYMOUS, resource, "get");
         // and being processed
         verify(requestMessageHandler).apply(any(Message.class), any(ResourceIdentifier.class));