+++ tests/src/test/java/org/eclipse/hono/tests/amqp/AmqpUploadTestBase.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2016, 2021 Contributors to the Eclipse Foundation
+ * Copyright (c) 2016, 2022 Contributors to the Eclipse Foundation
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -35,13 +35,13 @@ import org.eclipse.hono.application.client.DownstreamMessage;
 import org.eclipse.hono.application.client.MessageConsumer;
 import org.eclipse.hono.application.client.MessageContext;
 import org.eclipse.hono.client.ServerErrorException;
+import org.eclipse.hono.client.amqp.connection.AmqpConstants;
 import org.eclipse.hono.client.amqp.connection.AmqpErrorException;
 import org.eclipse.hono.service.management.device.Device;
 import org.eclipse.hono.service.management.tenant.Tenant;
 import org.eclipse.hono.tests.DownstreamMessageAssertions;
 import org.eclipse.hono.tests.IntegrationTestSupport;
 import org.eclipse.hono.tests.Tenants;
-import org.eclipse.hono.util.Constants;
 import org.eclipse.hono.util.EventConstants;
 import org.eclipse.hono.util.MessageHelper;
 import org.eclipse.hono.util.QoS;
@@ -169,7 +169,7 @@ public abstract class AmqpUploadTestBase extends AmqpAdapterTestBase {
                 assertThat(delivery.getRemoteState()).isInstanceOf(Rejected.class);
                 final Rejected rejected = (Rejected) delivery.getRemoteState();
                 final ErrorCondition error = rejected.getError();
-                assertThat((Object) error.getCondition()).isEqualTo(Constants.AMQP_BAD_REQUEST);
+                assertThat((Object) error.getCondition()).isEqualTo(AmqpConstants.AMQP_BAD_REQUEST);
             });
             context.completeNow();
         });