+++ site/homepage/content/release-notes.md
@@ -55,6 +55,8 @@ description = "Information about changes in recent Hono releases. Includes new f
 * Moved `org.eclipse.hono.util.TimeUntilDisconnectNotification` to application client module.
 * The `org.eclipse.hono.util.RequestResponseResult` class hierarchy's dependency on the Proton-J `ApplicationProperties` class
   has been removed. Instead, the constructors now accept a standard Map containing the properties.
+* Moved several AMQP 1.0 specific constants from `org.eclipse.hono.util.Constants` to
+  `org.eclipse.hono.client.amqp.connection.AmqpConstants` in the amqp-connection client module.
 
 ## 1.12.1